+++ lib/internal/encoding/mod.rs
@@ -1,4 +1,5 @@
 //! Functions having to do with encoding, decoding, and hashing data.
+//! Includes base64, sha1, and url encoding.
 
 pub mod base64;
 pub mod sha1;