+++ lib/internal/mod.rs
@@ -1,4 +1,4 @@
-//! Internal Functions
+//! Internal Functions.
 //!
 //! These may be helpful but may not remain compatible between updates.
 //! No guarantees are made about their stability, they are likely to be changed or removed.